Moved read_catalog into a more fitting io.py routine
