+++ Corrfunc/__init__.py
@@ -6,8 +6,8 @@ from __future__ import (division, print_function, absolute_import,
 import sys
 import os
 
-__all__ = ["_countpairs", "_countpairs_mocks", "utils", "theory", "mocks",
-           "tests"]
+__all__ = ["_countpairs", "_countpairs_mocks", "utils", "io",
+           "theory", "mocks", "tests"]
 
 # from Corrfunc import * throws: TypeError: Item in ``from list'' not a string
 # following the accepted answer in:
