+++ docs/source/modules/utilities.rst
@@ -6,7 +6,16 @@ Utility Routines
 All of these can be imported from `Corrfunc.utils`.
 
 .. currentmodule:: Corrfunc.utils
+.. autofunction:: translate_isa_string_to_enum
+.. autofunction:: return_file_with_rbins
+   
+                   
+Input/Output Routines
+======================
+All of the following functions can be imported from `Corrfunc.io`.
 
+.. currentmodule:: Corrfunc.io
+                   
 Reading input data
 ------------------
